src/requirements.txt
Django>=4.1.0,<4.2.0
python-decouple
psycopg2
gunicorn